Make response factoring clearer
