+++ src/OmniSharp.Abstractions/Models/v1/FileOperation/FileModificationType.cs
@@ -1,0 +1,9 @@
+﻿namespace OmniSharp.Models
+{
+    public enum FileModificationType
+    {
+        Modified,
+        Opened,
+        Renamed
+    }
+}
