+++ src/OmniSharp.Abstractions/Models/v1/FileOperation/FileOperationResponse.cs
@@ -1,0 +1,19 @@
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace OmniSharp.Models
+{
+    public abstract class FileOperationResponse
+    {
+        public FileOperationResponse(string fileName, FileModificationType type)
+        {
+            FileName = fileName;
+            ModificationType = type;
+        }
+
+        public string FileName { get; }
+
+        [JsonConverter(typeof(StringEnumConverter))]
+        public FileModificationType ModificationType { get; }
+    }
+}
