+++ src/OmniSharp.Abstractions/Models/v1/FileOperation/ModifiedFileResponse.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+
+namespace OmniSharp.Models
+{
+    public class ModifiedFileResponse : FileOperationResponse
+    {
+        public ModifiedFileResponse(string fileName)
+            : base(fileName, FileModificationType.Modified)
+        {
+        }
+
+        public string Buffer { get; set; }
+        public IEnumerable<LinePositionSpanTextChange> Changes { get; set; }
+    }
+}
