+++ src/OmniSharp.Abstractions/Models/v1/FileOperation/OpenFileResponse.cs
@@ -1,0 +1,9 @@
+﻿namespace OmniSharp.Models
+{
+    public class OpenFileResponse : FileOperationResponse
+    {
+        public OpenFileResponse(string fileName) : base(fileName, FileModificationType.Opened)
+        {
+        }
+    }
+}
