+++ src/OmniSharp.Abstractions/Models/v1/FileOperation/RenamedFileResponse.cs
@@ -1,0 +1,13 @@
+﻿namespace OmniSharp.Models
+{
+    public class RenamedFileResponse : FileOperationResponse
+    {
+        public RenamedFileResponse(string fileName, string newFileName)
+            : base(fileName, FileModificationType.Renamed)
+        {
+            NewFileName = newFileName;
+        }
+
+        public string NewFileName { get; }
+    }
+}
