+++ src/OmniSharp.Abstractions/Models/v2/RunCodeActionResponse.cs
@@ -4,6 +4,6 @@ namespace OmniSharp.Models.V2
 {
     public class RunCodeActionResponse
     {
-        public IEnumerable<ModifiedFileResponse> Changes { get; set; }
+        public IEnumerable<FileOperationResponse> Changes { get; set; }
     }
 }
